Avoid error when deleting an index
In the wait loop when deleting an index via the CLI, don't throw
an exception if the index is still in the ACTIVE state. There
is a small lapse in time from an update to the index state and the
time that the model is actually updated via ZK, so the ACTIVE state
should be treated in the same way as DELETE_REQUESTED in the wait
loop.