CDH-71906 IndexerIt#testLifecycleListeners fails with HBase 2.1 rebase
Adding a replica in HBase got slower and by the time the thread
is available the onSubscribe is not yet called causing mock
verification to fail.
By adding a timeout for verify the test passes, although it still
depends on thread name which is an internal implemetation of HBase
and could change in the future.

Change-Id: I802b2fc8a1528c3bae5d631fdf770f8138a72116